requirements.txt
streamlit
pandas
imapclient
pdfplumber
pyzmail36
dotenv
groq